Airflow101LearningPath/ca-demo-DAGs/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake==5.7.0
apache-airflow-providers-common-sql==1.16.0